+++ websites/rushstack.io/src/pages/community/profile.tsx
@@ -2,13 +2,13 @@ import React from "react";
 
 import { CommunitySidebarLayout } from "../../rscommunity/view/CommunitySidebarLayout";
 import { CommunitySignInPage } from "../../rscommunity/view/CommunitySignInPage";
-import { AppSession } from "../../rscommunity/model/AppSession";
+import { AppSession } from "../../rscommunity/api/AppSession";
 import {
   ApiTask,
   ApiTaskStatus,
   UserModel,
-} from "../../rscommunity/model/ApiDataService";
-import { IApiUser } from "../../rscommunity/ApiInterfaces";
+} from "../../rscommunity/api/ApiDataService";
+import { IApiUser } from "../../rscommunity/api/ApiInterfaces";
 import { ObjectEvent } from "../../rscommunity/library/ObjectEvent";
 import { DecoratedButton } from "../../rscommunity/view/DecoratedButton";
 import { FormFieldSet } from "../../rscommunity/form/FormFieldSet";
