+++ websites/rushstack.io/src/rscommunity/api/AppSession.ts
