+++ websites/rushstack.io/src/rscommunity/view/CommunitySidebarLayout.tsx
@@ -1,6 +1,6 @@
 import Layout from "@theme/Layout";
 import React from "react";
-import { AppSession } from "../model/AppSession";
+import { AppSession } from "../api/AppSession";
 
 interface IMenuItemProps {
   title: string;
